# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Collector.shop — API
  version: 0.1.0
  description: |
    Contrat des API HTTP du POC (catalogue-service et notification-service), écrit **avant** le code
    (règle A2-1). JSON en `snake_case`. Erreurs au format RFC 9457 (`application/problem+json`) avec
    une propriété `code` stable ; le détail technique (SQL, pile) reste dans les journaux.

    Trois user stories liées : **US-014** (mise en vente avec contrôle automatisé), **US-033**
    (revue des articles par l'administrateur), **US-029** (suivre un article, être prévenu des
    variations de prix). **US-021** (centres d'intérêt) est livrée en exemple guidé.

    Rôles Keycloak (`realm_access.roles`) : `acheteur`, `vendeur`, `admin`. Jetons de 5 minutes,
    RS256. Audience attendue : `catalogue-api` pour le catalogue, `notification-api` pour les notifications.
servers:
  - url: https://api.collector.local/api/v1
    description: Recette et démonstration (Traefik, TLS) — catalogue et notifications derrière le même hôte
  - url: http://localhost:8080/api/v1
    description: catalogue-service en local (docker compose)
  - url: http://localhost:8082/api/v1
    description: notification-service en local (docker compose)
...
paths:
...
  /me/notifications:
    get:
      tags: [acheteur]
      operationId: listNotifications
      summary: Mes notifications (US-029, CA-3)
      description: |
        Servi par **notification-service**. Un membre ne voit jamais les notifications des autres :
        l'identité est toujours celle du jeton (`sub`). Audience `notification-api`.
      security: [{ bearerAuth: [] }]
      servers:
        - url: https://api.collector.local/api/v1
      parameters:
        - name: unread
          in: query
          description: Seulement les non lues
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Notifications, les plus récentes d'abord
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Notification" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /me/notifications/{id}/read:
    post:
      tags: [acheteur]
      operationId: markNotificationRead
      summary: Marquer une notification lue (US-029, CA-5)
      description: Idempotent. La notification d'un autre membre répond 404, pas 403.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Marquée lue }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton OIDC Keycloak (Authorization Code + PKCE), RS256, 5 minutes.
...
  responses:
    InvalidRequest:
      description: "`invalid_request` — corps ou paramètre invalide ; `fields` liste les champs en erreur, jamais les valeurs saisies"
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: "`unauthorized` — jeton absent ou invalide"
    Forbidden:
      description: "`forbidden` — rôle manquant, ou article d'un autre vendeur"
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: "`not_found` — ressource inconnue, ou non visible pour l'appelant"
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      description: RFC 9457 avec un `code` stable (le message humain peut changer, le code non).
      required: [status, code]
      properties:
        type: { type: string, example: about:blank }
        title: { type: string }
        status: { type: integer, example: 422 }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          enum:
            - invalid_request
            - unauthorized
            - forbidden
            - not_found
            - invalid_status
            - too_many_photos
            - contact_info_forbidden
            - photo_required
            - invalid_photo
            - unknown_category
            - too_many_interests
            - internal_error
        fields:
          type: array
          description: Présent pour `invalid_request` de validation de corps
          items: { type: string }
...
    Notification:
      type: object
      required: [id, type, article_id, article_title, old_price_cents, new_price_cents, currency, price_drop, created_at]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [PRICE_CHANGED] }
        article_id: { type: string, format: uuid }
        article_title: { type: string }
        old_price_cents: { type: integer, format: int64 }
        new_price_cents: { type: integer, format: int64 }
        currency: { type: string, example: EUR }
        price_drop: { type: boolean }
        created_at: { type: string, format: date-time }
        read_at: { type: [string, "null"], format: date-time }